put runner.py into its own file instead of being a string in python.ts
